Added INI files to the lexers, enabled selecting more file types for packages, resolved a rare bug with the package manager, made it possible to pass the argument to "help" as a string

+++ kernel/core/strings/stringfunctions.hpp
@@ -444,7 +444,7 @@ static StringVector strfnc_ascii(StringFuncArgs& funcArgs)
 
     for (size_t i = 0; i < sView.length(); i++)
     {
-        sCodes.push_back((int)sView[i]);
+        sCodes.push_back((uint8_t)sView[i]);
     }
 
     return sCodes;
